backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
python-multipart==0.0.6
python-dotenv==1.0.0
argon2-cffi==23.1.0
passlib[argon2]==1.7.4
pydantic==2.5.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx[http2]==0.24.1